feat(heatmap): rail surface layer reusing the line-source kernel (M2)
Rail is an ISO 9613-2 line source with propagation identical to road
(cylindrical divergence + finite-line correction + terrain/screening/veg/
ground_g/reflection, barrier=max-not-sum); only emission (CNOSSOS Annex IV
rolling+traction), source height (0.5 m), and reach (7 km) differ — all
carried on the row. So M2 reuses the M1 scatter_line kernel UNCHANGED by
generalizing its row RoadRow -> LineRow, shared by both loaders via the new
source_line module.

source_loader_rail bakes per-period emission via normalize_rail ->
period_emissions — the exact chain the popup runs at RailSegment
construction — so the scatter matches compute_railways to quantisation noise.
The reach cull in compute_railways moves from a magic 8000 m to the named
RAILWAY_MAX_RADIUS: the popup already queries rail only within that radius,
so the old cull was dead code, and this aligns heatmap/oracle/popup exactly.

Also fixes a latent popup bug the rail parity review surfaced:
RailSegment.speed_kmh was u8, so high-speed rail's 300 km/h default
saturated to 255 (~1.4 dB too quiet) in both popup emission and display.
Widened to f64 (the emission model already supports 300 km/h); the heatmap
loader was already correct — this makes the popup oracle match it.

Parity GATE: validate-vs-popup --source rail, central-Praha tile vs
rail_periods -> 1000/1000 within +/-2 dB (p50 0.00, p95 0.23). Build clean
(0 warnings); 241 noise-compute + 38 heatmap-aircraft tests pass.

+++ engine/source-reader/src/lib.rs
@@ -214,7 +214,7 @@ pub fn collect_from_hex_data(
                 maxspeed: r.maxspeed,
                 trains_passenger: norm.scaled_passenger_per_day,
                 trains_freight: norm.scaled_freight_per_day,
-                speed_kmh: norm.speed_kmh as u8,
+                speed_kmh: norm.speed_kmh,
                 track_count: 1,
                 name: r.name.clone(),
                 rail_ref: r.rail_ref.clone(),
